Force 64-bit type on BITMASK [extended tests] [full tests]
On certain platforms, "unsigned long" is 32-bit wide, so (1UL << x) can
overflow if x is greater than 31, such as with HQC-5.

+++ src/kem/hqc/xuganyu96-hqc_hqc-3_ref/parameters.h
@@ -7,9 +7,10 @@
 #define HQC_PARAMETERS_H
 
 #include "api.h"
+#include <stdint.h>
 
 #define CEIL_DIVIDE(a, b) (((a) / (b)) + ((a) % (b) == 0 ? 0 : 1)) /*!< Divide a by b and ceil the result*/
-#define BITMASK(a, size)  ((1UL << (a % size)) - 1)                /*!< Create a mask*/
+#define BITMASK(a, size)  ((((uint64_t) 1UL) << (a % size)) - 1)   /*!< Create a mask*/
 
 #define PARAM_N                     35851       ///< Define the parameter n of the scheme
 #define PARAM_N1                    56          ///< Define the parameter n1 of the scheme (length of Reed-Solomon code)
